% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{clean_data}
\alias{clean_data}
\title{Clean and merge gambling premises data with geographical data}
\usage{
clean_data(gc_data, geo_data)
}
\arguments{
\item{gc_data}{Data frame containing gambling premises data}

\item{geo_data}{Data frame containing geographical data}
}
\value{
A merged data frame with cleaned and relevant columns)
}
\description{
Clean and merge gambling premises data with geographical data
}
