% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_df}
\alias{create_df}
\title{Create a data frame from clustering results and spatial data}
\usage{
create_df(res, xy)
}
\arguments{
\item{res}{Clustering results}

\item{xy}{Spatial data (points)}
}
\value{
A data frame with coordinates and cluster information
}
\description{
Create a data frame from clustering results and spatial data
}
