% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{create_df_null}
\alias{create_df_null}
\title{create a data frame containing the unnclustered points}
\usage{
create_df_null(df)
}
\arguments{
\item{df}{Data frame containing clustered points with columns LONG, LAT, cluster, ID}
}
\value{
A data frame with unclustered points (cluster == 0)
}
\description{
create a data frame containing the unnclustered points
}
