% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_cluster_hulls}
\alias{get_cluster_hulls}
\title{Get convex hulls for each cluster}
\usage{
get_cluster_hulls(df)
}
\arguments{
\item{df}{Data frame containing clustered points with columns LONG, LAT, cluster}
}
\value{
A list of data frames, each containing the convex hull for a cluster
}
\description{
Get convex hulls for each cluster
}
