% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{leaflet_map}
\alias{leaflet_map}
\title{Create a Leaflet map with cluster hulls and points}
\usage{
leaflet_map(cluster_hulls, df, df_null)
}
\arguments{
\item{cluster_hulls}{A list of data frames, each containing the convex hull for a cluster}

\item{df}{Data frame containing clustered points with columns LONG, LAT, cluster}

\item{df_null}{Data frame containing unclustered points with columns LONG, LAT, ID}
}
\value{
A leaflet map object
}
\description{
Create a Leaflet map with cluster hulls and points
}
